+++ docs/components/TheBreadcrumb.vue
@@ -1,0 +1,11 @@
+<template>
+    <nav class="breadcrumb" aria-label="breadcrumbs">
+        <ul>
+            <li><a href="#">Home</a></li>
+            <li><a href="#">Documentation</a></li>
+            <li class="is-active">
+                <a href="#" aria-current="page">Breadcrumb</a>
+            </li>
+        </ul>
+    </nav>
+</template>