+++ docs/pages/components/collapse/Collapse.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="default-container">
+    <div>
         <h1 class="title is-spaced">Collapse</h1>
         <h2 class="subtitle">Collapse is a easy way to toggle what you want</h2>
         <hr>