+++ docs/pages/components/tooltip/Tooltip.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="default-container">
+    <div>
         <h1 class="title is-spaced">Tooltip</h1>
         <h2 class="subtitle">Display a brief helper text to your user</h2>
         <hr>