+++ src/components/table/TableColumn.vue
@@ -39,7 +39,7 @@
                 }
             }
         },
-        created() {
+        beforeMount() {
             if (!this.$parent.$data._isTable) {
                 this.$destroy()
                 throw new Error('You should wrap bTableColumn on a bTable')
@@ -51,6 +51,12 @@
             // this finds when to stop based on the newKey property.
             const repeated = this.$parent.columns.some((column) => column.newKey === this.newKey)
             !repeated && this.$parent.columns.push(this)
+        },
+        beforeDestroy() {
+            const index = this.$parent.columns.map((column) => column.newKey).indexOf(this.newKey)
+            if (index >= 0) {
+                this.$parent.columns.splice(index, 1)
+            }
         }
     }
 </script>